requirements.txt.txt
flask
flask-cors
google-generativeai
google-api-python-client
pypdf